+++ src/main/java/com/kboticket/dto/UserDto.java
@@ -1,0 +1,11 @@
+package com.kboticket.dto;
+
+import lombok.Builder;
+import lombok.Getter;
+
+@Builder
+@Getter
+public class UserDto {
+    private String email;
+    private String phone;
+}
