+++ src/main/java/com/kboticket/repository/TermsRepository.java
@@ -2,6 +2,7 @@ package com.kboticket.repository;
 
 import com.kboticket.domain.Terms;
 import com.kboticket.domain.TermsPk;
+import com.kboticket.enums.TermsType;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.querydsl.QuerydslPredicateExecutor;
 import org.springframework.stereotype.Repository;
@@ -12,6 +13,5 @@ import java.util.List;
 @Repository
 public interface TermsRepository extends JpaRepository<Terms, TermsPk>, QuerydslPredicateExecutor<Terms> {
 
-    List<Terms> findByMandatoryTrue();
-
+    List<Terms> findByMandatoryTrueAndType(TermsType type);
 }
